+++ lib/home/view/images/random_bird.dart
@@ -28,4 +28,23 @@ class _RandomBirdState extends ImageAPIState<RandomBird> {
         ) {
     add(BirdController());
   }
+
+  /// Overridden for your benefit.
+  /// Place a breakpoint here and note when its called.
+  @override
+  // ignore: unnecessary_overrides
+  Widget build(BuildContext context) => super.build(context);
+
+  /// Overridden for your benefit.
+  /// Place a breakpoint here and note when its called.
+  /// Right after the initState() function when first created and
+  /// whenever its to rebuild due its possible dependency to an InheritedWidget
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    if (kDebugMode) {
+      print(
+          '>>>>>>>>>>>>>>>>>>>>>>>>>   _RandomBirdState didChangeDependencies() called.');
+    }
+  }
 }