+++ lib/home/view/images/random_cat.dart
@@ -34,4 +34,23 @@ class _RandomCatState extends ImageAPIState<RandomCat> {
         ) {
     add(CatController());
   }
+
+  /// Overridden for your benefit.
+  /// Place a breakpoint here and note when its called.
+  @override
+  // ignore: unnecessary_overrides
+  Widget build(BuildContext context) => super.build(context);
+
+  /// Overridden for your benefit.
+  /// Place a breakpoint here and note when its called.
+  /// Right after the initState() function when first created and
+  /// whenever its to rebuild due its possible dependency to an InheritedWidget
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    if (kDebugMode) {
+      print(
+          '>>>>>>>>>>>>>>>>>>>>>>>>>   _RandomCatState didChangeDependencies() called.');
+    }
+  }
 }