+++ lib/home/view/mixin/_view_mixin.dart
@@ -1,4 +1,6 @@
 //
-export 'futurebuilder_mixin.dart';
+//
+//export 'futurebuilder_mixin.dart';
+export 'futurebuilder_state_mixin.dart';
 
 export 'inheritedwidget_state_mixin.dart';