+++ lib/src/view.dart
@@ -1,6 +1,8 @@
 //
 export 'package:flutter/material.dart';
 
+export 'package:flutter/foundation.dart' show kDebugMode;
+
 export '/app/view_app.dart';
 
 export '/home/view_home.dart';